requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
fuzzywuzzy==0.18.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Levenshtein==0.23.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-Levenshtein==0.23.0
rapidfuzz==3.5.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
twisted==23.10.0
txaio==23.1.1
typing-extensions==4.8.0
zope.interface==6.1
